TCG-110: Refactor - separate MtgjsonPricesV2Container into its own module
- Move MtgjsonPricesV2Container from mtgjson_prices_v2.py to mtgjson_prices_v2_container.py
- Better separation of concerns: record class and container class now in separate files
- Update imports in __init__.py to maintain backward compatibility
- All tests pass, mypy and pylint clean
